requirements.txt
amqp==5.3.1
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
babel==2.17.0
billiard==4.2.1
black==24.3.0
celery==5.3.6
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
dacite==1.9.2
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2.20
django-allauth==0.60.1
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-money==3.4.1
django-redis==5.4.0
django-storages==1.14.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
factory-boy==3.3.0
Faker==37.0.2
filelock==3.18.0
flake8==7.0.0
fonttools==4.56.0
fsspec==2025.3.0
ftfy==6.3.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.26.0
idna==3.10
iniconfig==2.1.0
isort==5.13.2
Jinja2==3.1.6
kiwisolver==1.4.8
kombu==5.5.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mccabe==0.7.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.1.1
oauthlib==3.2.2
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.3.7
pluggy==1.5.0
pokemontcgsdk==3.4.0
prompt_toolkit==3.0.50
psutil==7.0.0
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
py-moneyed==3.0
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.10.1
pyparsing==3.2.2
pytest==7.4.4
pytest-django==4.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2025.1
PyYAML==6.0.2
redis==5.0.1
regex==2024.11.6
requests==2.32.0
requests-oauthlib==2.0.0
scipy==1.15.2
seaborn==0.13.2
setuptools==77.0.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stripe==7.10.0
sympy==1.13.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.95
ultralytics-thop==2.0.14
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
